renamed base parallel env file
